internal/audio: fix ADPCM decoding due to wrong use of ints
Previously, the code used an int16 to store newSample. However, this
meant the clamp check was always bypassed because an int16 is always
going to be within the limits of min and max int16s.

There was always some popping and crackling in the generated .wav files
and this has fixed it. E.g. in 09EERIE.WAVE a pop sound could be heard
at around 58 seconds.

A hint for the fix was given by staticchecek which warned that an int16
was always going to be within the limits of min and max int16s.